create util function to transform EAP data into timeseries
