+++ static/app/views/performance/transactionSummary/transactionOverview/utils.tsx
@@ -1,8 +1,11 @@
 import type {Organization} from 'sentry/types/organization';
+import type {DataUnit} from 'sentry/utils/discover/fields';
 import type {MetricsCardinalityContext} from 'sentry/utils/performance/contexts/metricsCardinality';
 import type {MetricsEnhancedPerformanceDataContext} from 'sentry/utils/performance/contexts/metricsEnhancedPerformanceDataContext';
 import type {MetricsEnhancedSettingContext} from 'sentry/utils/performance/contexts/metricsEnhancedSetting';
 import {canUseMetricsData} from 'sentry/utils/performance/contexts/metricsEnhancedSetting';
+import type {TimeSeries} from 'sentry/views/dashboards/widgets/common/types';
+import type {DiscoverSeries} from 'sentry/views/insights/common/queries/useDiscoverSeries';
 import {getMEPQueryParams} from 'sentry/views/performance/landing/widgets/utils';
 
 export function canUseTransactionMetricsData(
@@ -36,4 +39,25 @@ export function getTransactionMEPParamsIfApplicable(
   }
 
   return getMEPQueryParams(mepSetting, true);
+}
+
+type EAPSeriesData = Record<string, DiscoverSeries>;
+export function eapSeriesDataToTimeSeries(data: EAPSeriesData) {
+  const timeSeries: TimeSeries[] = [];
+  Object.entries(data).forEach(([key, value]) => {
+    timeSeries.push({
+      field: key,
+      meta: {
+        type: value.meta?.fields?.[key] ?? null,
+        unit: value.meta?.units?.[key] as DataUnit,
+      },
+      data:
+        value.data.map(item => ({
+          timestamp: item.name.toString(),
+          value: item.value,
+        })) ?? [],
+    });
+  });
+
+  return timeSeries;
 }
